feat: :sparkles: Creating PersonModel class configures the database
This is a Java class that represents a person model and uses JPA (Java Persistence API) annotations to map the class to a database table. The class has fields for a person's ID, name, birthdate and a list of addresses. The JsonIdentityInfo annotation is used for handling circular dependencies when serializing the object to JSON using the Jackson library. The @Entity annotation indicates that this class is an entity and should be mapped to a database table. The @Id and @GeneratedValue annotations are used to indicate that the "id" field is the primary key and its value is generated automatically. The @Column annotation is used to specify the column name and constraints for each field in the table. The @OneToMany annotation is used to indicate a one-to-many relationship between the PersonModel and AddressModel classes, where one person can have multiple addresses.