feat: :sparkles: Creating AddressModel class that configures the database
This is a Java class that represents an Address model and uses JPA (Java Persistence API) annotations to map the class to a database table. The class has fields for an address's ID, city, main flag, number, public place, zip code and a person. The JsonIdentityInfo annotation is used for handling circular dependencies when serializing the object to JSON using the Jackson library. The @Entity annotation indicates that this class is an entity and should be mapped to a database table. The @Table annotation is used to specify the name of the database table that this class maps to. The @Id and @GeneratedValue annotations are used to indicate that the "id" field is the primary key and its value is generated automatically using a sequence generator. The @SequenceGenerator annotation is used to configure the sequence generator. The @ManyToOne annotation is used to indicate a many-to-one relationship between the AddressModel and PersonModel classes, where one person can have multiple addresses but each address belongs to only one person.